binlog_reader.[hc] completed

+++ src/server/storage/inode/binlog_reader.h
@@ -24,12 +24,12 @@
 extern "C" {
 #endif
 
-int binlog_reader_load(const int binlog_id,
+int binlog_reader_load(const uint64_t binlog_id,
         FDIRStorageInodeIndexArray *index_array);
 
-int binlog_reader_get_first_inode(const int binlog_id, int64_t *inode);
+int binlog_reader_get_first_inode(const uint64_t binlog_id, int64_t *inode);
 
-int binlog_reader_get_last_inode(const int binlog_id, int64_t *inode);
+int binlog_reader_get_last_inode(const uint64_t binlog_id, int64_t *inode);
 
 #ifdef __cplusplus
 }
